Task 1

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,5 +1,5 @@
 #include "function_pointers.h"
-#include <stdlib.h>
+#include <stddef.h>
 
 /**
  * array_iterator - executes function given as param. on each element of array
@@ -10,7 +10,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	int i;
+	size_t i;
 
 	for (i = 0; i < size; i++)
 	{
